+++ docs/iframe.html
@@ -130,4 +130,4 @@
             
         
             
-              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.77ba099e85a299c32745.bundle.js"></script><script src="vendors~main.77ba099e85a299c32745.bundle.js"></script><script src="main.77ba099e85a299c32745.bundle.js"></script></body></html>
+              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.bc3aa2bdfc9b79f4fbd4.bundle.js"></script><script src="vendors~main.bc3aa2bdfc9b79f4fbd4.bundle.js"></script><script src="main.bc3aa2bdfc9b79f4fbd4.bundle.js"></script></body></html>
