+++ docs/index.html
@@ -52,4 +52,4 @@
             
         
             
-              window['DOCS_MODE'] = false;</script><script src="runtime~main.870e8ee0658e83d05c6b.bundle.js"></script><script src="vendors~main.77d60a20c6099ead23b7.bundle.js"></script><script src="main.691c003d4ab0e90a3ff6.bundle.js"></script></body></html>
+              window['DOCS_MODE'] = false;</script><script src="runtime~main.870e8ee0658e83d05c6b.bundle.js"></script><script src="vendors~main.cf019410aa62abcf7914.bundle.js"></script><script src="main.594cd98c4eacb39a9f15.bundle.js"></script></body></html>
