+++ src/mixins/index.ts
@@ -1,11 +1,9 @@
 import CommonMixin from './common-mixin'
-import CurrencyLookupMixin from './currency-lookup-mixin'
 import DateMixin from './date-mixin'
 import FilingMixin from './filing-mixin'
 
 export {
   CommonMixin,
-  CurrencyLookupMixin,
   DateMixin,
   FilingMixin
 }
